requirements.txt
anthropic
Pillow
pillow-heif
pdf2image
